chore(debug): refactor types
